requirements.txt
Pillow==9.1.1
pyserial==3.5
PyYAML~=6.0
psutil~=5.9.1